refactor(studio): stop workspace routes loading legacy Event Studio JS
Route-aware library attachment keeps mel_event_studio on legacy wizard
routes only; workspace sections get shell plus section-specific bundles
(AI assist, branding, ticket preview). Extracts AI assist into its own
library and adds unit contract tests.
